Remove OpenCensus receiver, add OpenTelemetry receiver
