+++ oap-server/server-receiver-plugin/otel-receiver-plugin/pom.xml
@@ -25,7 +25,7 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>opencensus-receiver-plugin</artifactId>
+    <artifactId>otel-receiver-plugin</artifactId>
     <packaging>jar</packaging>
 
     <dependencies>
