+++ oap-server/server-receiver-plugin/otel-receiver-plugin/src/main/resources/META-INF/services/org.apache.skywalking.oap.server.library.module.ModuleDefine
@@ -17,4 +17,4 @@
 #
 
 
-org.apache.skywalking.oap.server.receiver.opencensus.OCMetricReceiverModule
+org.apache.skywalking.oap.server.receiver.otel.OtelMetricReceiverModule
