+++ oap-server/server-receiver-plugin/otel-receiver-plugin/src/main/resources/META-INF/services/org.apache.skywalking.oap.server.library.module.ModuleProvider
@@ -16,4 +16,4 @@
 #
 #
 
-org.apache.skywalking.oap.server.receiver.opencensus.OCMetricReceiverProvider
+org.apache.skywalking.oap.server.receiver.otel.oc.OCMetricReceiverProvider
