+++ oap-server/server-receiver-plugin/pom.xml
@@ -40,7 +40,7 @@
         <module>jaeger-receiver-plugin</module>
         <module>receiver-proto</module>
         <module>skywalking-profile-receiver-plugin</module>
-        <module>opencensus-receiver-plugin</module>
+        <module>otel-receiver-plugin</module>
         <module>skywalking-meter-receiver-plugin</module>
         <module>skywalking-browser-receiver-plugin</module>
     </modules>
